Removido arquivos desnecessários
