--- dev/.vscode/settings.json
@@ -1,34 +1,0 @@
-{
-  "editor.formatOnPaste": false,
-  "editor.formatOnSave": false,
-  "editor.formatOnType": false,
-  "eslint.validate": ["javascript", "javascriptreact", "typescript", "typescriptreact"],
-  "files.exclude": {
-    "build": true,
-    "**/*.css": {
-      "when": "$(basename).scss"
-    },
-    "**/*.js": {
-      "when": "$(basename).ts"
-    },
-    "**/*?.js": {
-      "when": "$(basename).tsx"
-    },
-    "**/*.lock.json": true,
-    "**/*.map": {
-      "when": "$(basename)"
-    },
-    "**/.DS_Store": true,
-    "**/.git": true,
-    "**/.svn": true,
-    "**/.tmp": true,
-    "**/bin/**": true,
-    "**/obj/**": true
-  },
-  "typescript.tsdk": "node_modules/typescript/lib",
-  // "typescriptHero.imports.organizeOnSave": true,
-  "typescriptHero.imports.multiLineTrailingComma": false,
-  "editor.codeActionsOnSave": {
-    "source.fixAll.eslint": true
-  }
-}
