+++ dev/package.json
@@ -39,7 +39,6 @@
     "raw-loader": "4.0.2",
     "react-app-rewired": "2.1.6",
     "source-map-explorer": "2.5.0",
-    "typescript": "4.0.5",
     "webpack-bundle-analyzer": "4.1.0"
   },
   "resolutions": {
