+++ main.code-workspace
@@ -1,46 +1,81 @@
 {
-	"folders": [
-		{
-			"name": "all",
-			"path": "."
-		},
-		{
-			"path": "./dev"
-		},
-		{
-			"path": "./docs"
-		},
-		{
-			"name": "@eduzz/core",
-			"path": "./packages/core"
-		},
-		{
-			"name": "@eduzz/houston-ui",
-			"path": "./packages/ui-components"
-		},
-		{
-			"name": "@eduzz/houston-forms",
-			"path": "./packages/forms"
-		},
-		{
-			"name": "@eduzz/eslint-config-houston",
-			"path": "./packages/eslint-config"
-		},
-		{
-			"name": "@eduzz/houston",
-			"path": "./packages/bundle"
-		}
-	],
-	"settings": {
-		"typescript.tsdk": "./node_modules/typescript/lib"
-	},
-	"extensions": {
-		"recommendations": [
-			"darkriszty.markdown-table-prettify"
-		]
-	},
-	"recommendations": [
-		"darkriszty.markdown-table-prettify",
-		"dbaeumer.vscode-eslint"
-	]
+  "folders": [
+    {
+      "name": "all",
+      "path": "."
+    },
+    {
+      "path": "./dev"
+    },
+    {
+      "path": "./docs"
+    },
+    {
+      "name": "@eduzz/core",
+      "path": "./packages/core"
+    },
+    {
+      "name": "@eduzz/houston-ui",
+      "path": "./packages/ui-components"
+    },
+    {
+      "name": "@eduzz/houston-forms",
+      "path": "./packages/forms"
+    },
+    {
+      "name": "@eduzz/eslint-config-houston",
+      "path": "./packages/eslint-config"
+    },
+    {
+      "name": "@eduzz/houston",
+      "path": "./packages/bundle"
+    }
+  ],
+  "settings": {
+    "typescript.tsdk": "./node_modules/typescript/lib",
+    "editor.formatOnPaste": false,
+    "editor.formatOnSave": false,
+    "editor.formatOnType": false,
+    "eslint.validate": [
+      "javascript",
+      "javascriptreact",
+      "typescript",
+      "typescriptreact"
+    ],
+    "files.exclude": {
+			"node_modules": true,
+			".ultra.cache.json": true,
+      "**/*.css": {
+        "when": "$(basename).scss"
+      },
+      "**/*.js": {
+        "when": "$(basename).ts"
+			},
+			"**/*.d.ts": {
+        "when": "$(basename).ts"
+      },
+      "**/*?.js": {
+        "when": "$(basename).tsx"
+      },
+      "**/*.lock.json": true,
+      "**/*.map": {
+        "when": "$(basename)"
+      },
+      "**/.DS_Store": true,
+      "**/.git": true,
+      "**/.svn": true,
+      "**/.tmp": true,
+      "**/bin/**": true,
+      "**/obj/**": true
+    },
+    "editor.codeActionsOnSave": {
+      "source.fixAll.eslint": true
+    }
+  },
+  "extensions": {
+    "recommendations": [
+      "darkriszty.markdown-table-prettify",
+      "dbaeumer.vscode-eslint"
+    ]
+  }
 }
