+++ package.json
@@ -19,6 +19,7 @@
   },
   "devDependencies": {
     "husky": "4.3.0",
+    "typescript": "4.1.3",
     "ultra-runner": "3.7.1"
   },
   "ultra": {
