--- packages/eslint-config/.vscode/extensions.json
@@ -1,6 +1,0 @@
-{
-  "recommendations": [
-    "darkriszty.markdown-table-prettify",
-    "dbaeumer.vscode-eslint"
-  ]
-}
