+++ packages/eslint-config/package.json
@@ -25,8 +25,7 @@
     "eslint-plugin-react": "^7",
     "eslint-plugin-react-hooks": "^4",
     "eslint-plugin-unused-imports": "^1",
-    "prettier": "^2",
-    "typescript": "^4"
+    "prettier": "^2"
   },
   "devDependencies": {},
   "peerDependencies": {
@@ -40,7 +39,6 @@
     "eslint-plugin-react": "^7",
     "eslint-plugin-react-hooks": "^4",
     "eslint-plugin-unused-imports": "^1",
-    "prettier": "^2",
-    "typescript": "^4"
+    "prettier": "^2"
   }
 }
