+++ packages/forms/package.json
@@ -38,8 +38,7 @@
   },
   "devDependencies": {
     "@eduzz/eslint-config-houston": "*",
-    "@types/react": "^17.0.0",
-    "concurrently": "5.3.0",
-    "typescript": "4.0.5"
+    "@types/react": "16.9.55",
+    "concurrently": "5.3.0"
   }
 }
