--- packages/ui-components/.vscode/settings.json
@@ -1,32 +1,0 @@
-{
-  "eslint.validate": [
-    "javascript",
-    "javascriptreact",
-    "typescript",
-    "typescriptreact"
-  ],
-  "files.exclude": {
-    "**/*.lock.json": true,
-    "**/.DS_Store": true,
-    "**/.git": true,
-    "**/.svn": true,
-    "**/.tmp": true,
-    "**/bin/**": true,
-    "**/obj/**": true,
-    "node_modules": true,
-    "**/**/*.js": {
-      "when": "$(basename).ts"
-    },
-    "**/**/**.js": {
-      "when": "$(basename).tsx"
-    },
-    "**/**/*.d.ts": true
-  },
-  "typescript.tsdk": "./node_modules/typescript/lib",
-  "typescriptHero.imports.multiLineTrailingComma": false,
-  "typescriptHero.imports.organizeOnSave": false,
-  "editor.formatOnSave": false,
-  "editor.codeActionsOnSave": {
-    "source.fixAll.eslint": true
-  }
-}
