+++ packages/ui-components/README.md
@@ -1,10 +1,7 @@
 # Eduzz Houston
 
-<img src="docs/assets/welcome.svg" width="100%" />
-
-[![version](https://img.shields.io/npm/v/@eduzz/houston?color=%23E3F2FF)](https://www.npmjs.com/package/@eduzz/houston)
-[![size](https://img.shields.io/bundlephobia/min/@eduzz/houston?color=%23E3F2FF)](https://www.npmjs.com/package/@eduzz/houston)
-[![version](https://img.shields.io/github/last-commit/eduzz/houston?color=%23E3F2FF)](https://github.com/eduzz/houston/commits)
+[![version](https://img.shields.io/npm/v/@eduzz/@eduzz/houston-ui)](https://www.npmjs.com/package/@eduzz/houston)
+[![size](https://img.shields.io/bundlephobia/min/@eduzz/houston-ui)](https://www.npmjs.com/package/@eduzz/houston)
 
 ### Installation
