+++ packages/ui-components/ThemeProvider/index.tsx
@@ -3,12 +3,9 @@ import * as React from 'react';
 import { Theme, ThemeProviderProps } from '@material-ui/core/styles';
 import { PaletteOptions } from '@material-ui/core/styles/createPalette';
 
-import { ToastContainer } from '@eduzz/houston-ui/Toast';
-
+import ToastContainer from '../Toast/Container';
 import ContextTheme from './context';
 import generateCustomTheme from './custom';
-
-import 'react-toastify/dist/ReactToastify.css';
 
 export type IThemeProviderApplications = 'select' | 'orbita' | 'blinket' | 'backoffice' | 'custom';
